OMCPT-40: added make command to generate step documentation
A python script has been added that scans a folder
and creates markdown files for each feature file and for
each python implementation of gherkin steps.
The script adds a folder, populates it with markdown files
and an index.md.
This folder is added to git so that it can be accessed via gitlab.